fix: address unresolved review threads from merged PRs
1. Add kani::cover! vacuity guard to CRI no-panic proof (#1789 thread)
   - verify_parse_cri_line_no_panic now asserts the Some path is
     reachable, preventing vacuous proofs.

2. Consolidate parse_content_type into shared receiver_http module
   (#1797 thread) — eliminates 3 identical copies across
   arrow_ipc_receiver, otap_receiver, and otlp_receiver/server.
