+++ crates/logfwd-io/src/otlp_receiver/server.rs
@@ -11,7 +11,7 @@ use axum::response::{IntoResponse, Response};
 use logfwd_types::diagnostics::{ComponentHealth, ComponentStats};
 
 use crate::InputError;
-use crate::receiver_http::{MAX_REQUEST_BODY_SIZE, declared_content_length, read_limited_body};
+use crate::receiver_http::{MAX_REQUEST_BODY_SIZE, declared_content_length, parse_content_type, read_limited_body};
 
 use super::decode::{decode_otlp_json, decode_otlp_protobuf, decompress_gzip, decompress_zstd};
 use super::{OtlpServerState, ReceiverPayload};
@@ -180,14 +180,3 @@ fn parse_content_encoding(headers: &HeaderMap) -> Result<Option<String>, StatusC
     Ok(Some(parsed.to_ascii_lowercase()))
 }
 
-fn parse_content_type(headers: &HeaderMap) -> Result<Option<String>, StatusCode> {
-    let Some(value) = headers.get(CONTENT_TYPE) else {
-        return Ok(None);
-    };
-    let raw = value.to_str().map_err(|_| StatusCode::BAD_REQUEST)?;
-    let media_type = raw.split(';').next().unwrap_or_default().trim();
-    if media_type.is_empty() {
-        return Err(StatusCode::BAD_REQUEST);
-    }
-    Ok(Some(media_type.to_ascii_lowercase()))
-}
